move theming types into a subdirectory
