+++ packages/agentic-design/src/theming/context.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import type { Theme } from './Theme.types';
+import type { Theme } from './types/Theme.types';
 
 export const ThemeContext = React.createContext<Theme>(undefined);
 export const useTheme = () => React.useContext(ThemeContext);
